Session 4

+++ DemoProsjekt/Demo.cs
@@ -46,7 +46,7 @@ namespace DemoProsjekt
         {
             if (b == 0)
                 return 0;
-            return a / b;
+            return (double)a / (double)b;
         }
 
         public int GetDemoId()
